# cmd/snap: speed up unit tests