cmd/snap started at nearly 13s, but most of that is because ssh2 is
slow to start up. I was able to shave a couple of seconds elsewhere
though. And I tagged the ssh2 tests as not-short.

Before:

    $ go test -count 1 ./cmd/snap
    ok  	github.com/snapcore/snapd/cmd/snap	13.868s

and then,

    $ go test -count 1 ./cmd/snap
    ok  	github.com/snapcore/snapd/cmd/snap	11.470s
    $ go test -short -count 1 ./cmd/snap
    ok  	github.com/snapcore/snapd/cmd/snap	1.251s