@chipaca 2018-10-08
I'll bump the loop to 1000, tonight  probably (don't want this silly thing spamming travis during the day)